+++ src/app/auth/services/interfaces.ts
